test(adapter-openclaw): split DKG channel test suite
